Per-eye field of view, projection matrices for left and right eye. Closes GH-1
The projection matrix rarely changes, only needing to be updated when
the window is resized (updateProjectionMatrix event from voxel-shader),
not each rendering tick, but for VR, it differs between the left and
right eye. We still only compute the matrices when needed, but store
them in our own plugin and update voxel-shader .projectionMatrix for
each eye as needed in our rendering method.

Test the independent FOVs using something like this from the console:

    t=game.plugins.get('voxel-vr'); t.FOVsLeft.upDegrees += 1; t.shader.updateProjectionMatrix()