Fix updating projection matrix after enable even if no VR devices present
Normally the projection matrix is updated when the game shell is
resized, or if a VR device is found (with the (asynchronous) WebVR) APIs
and then its parameters can be read and settings configured
appropriately. If there are no devices, enabling the plugin would
render nothing until the window was resized, due to this missing call.